Refactored vargs
Changed parseArgs to create a new vargs object.
To get a map or object literal, there are new
methods of toMap() and toObjectLiteral().

Changed many methods to getter and setters
whenever possible.

Commented out old functions.

Removed type key on argument objects, now using
the presence or absence of the position key to
determine whether an argument is positional or
not.