Added help functionality and support for default argument values
Added setDefaultValues function to allow for options and
positionals with default values.

E.g. if an option or positional have the 'value' property,
aren't required, aren't a flag (options only), and do not appear
in the resulting `vargs` object following a `parse`, then a positional
or option with the default value set is added to the resultant `vargs`.

Added help getters, help property to Arg, etc.

Added tests for all the new features.

Signed-off-by: blueblots <63152708+blueblots@users.noreply.github.com>