Implemented multiple value options
Now supports `+`, `*`, `5 (could be any number)` values
for the new `nargs` option parameter.

`+` = one or more values, if there are none then the option
is not included in `parse` output; if the option has a default
value, then it returns the default value.

`*` = zero or more values, if there are no values then the
option is included as an empty array.

`5` etc. = five or more values, option will consume up to this
many values in the command string.

Signed-off-by: blueblots <63152708+blueblots@users.noreply.github.com>